moving addToSourceDense and updateExplicitQuantities_
to FlowProblemBlackoil
